Update APS.shift_next to reduce by replaying the non-terminal.
This matches the way Reduce actions are implemented in the Python and Rust
runtime. However, this requires an extra `reducing` flag to prevent infinite
loops of APS when matching empty productions with inconsistent parse tables.

This change is useful for the upcoming Replay action, as we would have to split
the action state which is being visited from the stack and shift states.