Add remove_reduce transformation.
This transformation convert Reduce(Unwind(...)) actions into its Unwind action
followed by FilterStates and Replay edges.

This transformation pre-compute the determinisitic path taken while reducing,
and adds it back to the graph. Before this transformation, the parser stack was
completely controlled by shift-ing edges and reducing-non-terminals. After this
transformation, the state of the Parser might be disconnected from the last
state pushed to the stack.