QueueStack.pop_n: Remove unneeded assertion.
Currently `pop_n` is only used when the gap is 0.

However, this assertion will not hold if we manage to skip calling `push_next`
and transfert computed value out of the stack. This might happen in the future
as we remove replayed terms, in which case multiple calls to QueueStack.pop_n
might follow each others. Thus, the gap might not always be 0.