Replace Term enum by a simple number.
This change replace the Term enum which was wrapping the enumerated type of
TerminalId and NonterminalId values. Despite the non-overlapping ranges of
numbers, Rust kept a single bit to distinguish between a Terminal and a
Nonterminal.

This change encode a Term as a single integer, as it was expected to be compiled
to. Note, this is using u32 instead of the minimal u16 representation as u32 is
faster to load and write to memory.